Lab06/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
numpy>=1.22
Pillow>=9.0.0
matplotlib>=3.6.0
pyyaml>=6.0
